AvielCo/issue49

+++ backend/py-files/crop.py
@@ -5,7 +5,7 @@ import cv2
 import numpy as np
 from PIL import Image
 from PIL import UnidentifiedImageError
-
+import json
 from consts import *
 from dual_print import dual_print
 
@@ -61,9 +61,9 @@ def crop_image_to_patches(bw_img, grayscale_img, image_width, image_height, imag
     if shape_type is not None:
         items_in_folder = len(os.listdir(os.path.join(OUTPUT_PATH, shape_type, folder_name)))
         if shape_type == "cursive" and items_in_folder >= 8000:
-            return False
+            raise Exception("Unknown error, please try again.")
         if shape_type != "cursive" and items_in_folder >= 4000:
-            return False
+            raise Exception("Unknown error, please try again.")
     global total_images_cropped
     global total_patches_cropped
     x1 = y1 = 0
@@ -87,9 +87,9 @@ def crop_image_to_patches(bw_img, grayscale_img, image_width, image_height, imag
                                              )  # save location: output\\shape_type\\folder_name\\image_name_i.jpg
 
             else:
-                save_location = os.path.join(PROJECT_DIR,
-                                             "prediction_patches",
-                                             "1",
+                save_location = os.path.join(PREDICT_DIR,
+                                             "predict_patches",
+                                             folder_name,
                                              image_name + "_" + str(i) + ".jpg")
 
             if is_good_patch(bw_cropped_patch):
@@ -102,7 +102,6 @@ def crop_image_to_patches(bw_img, grayscale_img, image_width, image_height, imag
 
     total_images_cropped += 1
     dual_print(f"Successfully cropped to patches {image_name} in {save_location}, with shape: {shape_type}")
-    return True
 
 
 def binarization(img):
@@ -118,12 +117,16 @@ def binarization(img):
     return img
 
 
-def crop_image_edges(image_path):
-    img_path = os.path.join(BUFFER_PATH, "buffer_img.jpg")
+def crop_image_edges(image_path, folder_name):
+    try:
+        os.makedirs(os.path.join(BUFFER_PATH, folder_name))
+    except FileExistsError:
+        pass
+    img_path = os.path.join(BUFFER_PATH, folder_name, "buffer_img.jpg")
     if os.path.exists(img_path):
         os.remove(img_path)
-    img_path_left = os.path.join(BUFFER_PATH, "buffer_img_1.jpg")
-    img_path_right = os.path.join(BUFFER_PATH, "buffer_img_2.jpg")
+    img_path_left = os.path.join(BUFFER_PATH, folder_name, "buffer_img_1.jpg")
+    img_path_right = os.path.join(BUFFER_PATH, folder_name, "buffer_img_2.jpg")
     try:
         os.remove(img_path_left)
     except FileNotFoundError:
@@ -148,7 +151,7 @@ def crop_image_edges(image_path):
     try:
         img = img.crop(coords)
     except OSError:
-        return False
+        raise Exception("Unknown error, please try again.")
 
     w, h = img.size
     ratio = h / w
@@ -161,19 +164,17 @@ def crop_image_edges(image_path):
             cv2.imwrite(img_path_left, cv2.cvtColor(left_side, cv2.COLOR_RGB2BGR))
             cv2.imwrite(img_path_right, cv2.cvtColor(right_side, cv2.COLOR_RGB2BGR))
         except cv2.error as e:
-            dual_print(e, "error")
-            return False
+            raise Exception("Unknown error, please try again.")
     else:  # page has only 1 side
         np_img = np.array(img)
         try:
             cv2.imwrite(img_path, cv2.cvtColor(np_img, cv2.COLOR_RGB2BGR))
         except cv2.error as e:
             dual_print(e, "error")
-            return False
-    return True
+            raise Exception("Unknown error, please try again.")
 
 
-def process_image(path: str, folder_name: str, image_name: str, shape_type=None):
+def process_image(path, folder_name, image_name, shape_type=None):
     """
     This function process an image
         1. remove the edges of the image
@@ -185,14 +186,16 @@ def process_image(path: str, folder_name: str, image_name: str, shape_type=None)
     folderName (str): The name of the folder that the scan is saved in.
     """
     full_img_path = os.path.join(path, folder_name, image_name)
-    t = True
-    if not crop_image_edges(full_img_path):
+    try:
+        crop_image_edges(full_img_path, folder_name)
+    except Exception as e:
         os.remove(full_img_path)
-        return t
-    for _, _, files in os.walk(BUFFER_PATH):
+        raise e
+    FULL_BUFFER_PATH = os.path.join(BUFFER_PATH, folder_name)
+    for _, _, files in os.walk(FULL_BUFFER_PATH):
         i = 0
         for file in files:
-            file_path = os.path.join(BUFFER_PATH, file)
+            file_path = os.path.join(FULL_BUFFER_PATH, file)
             grayscale_img = cv2.imread(file_path, 0)  # Read the image from the folder with grayscale mode
             image_name_no_extension = os.path.splitext(image_name)[0]  # For the log
             if i == 0:
@@ -203,13 +206,15 @@ def process_image(path: str, folder_name: str, image_name: str, shape_type=None)
             h, w = dims[0], dims[1]
             bw_img = binarization(grayscale_img)  # Open pic in BW
 
-            t = crop_image_to_patches(bw_img, grayscale_img, w, h, image_name_no_extension, folder_name, shape_type)
+            try:
+                crop_image_to_patches(bw_img, grayscale_img, w, h, image_name_no_extension, folder_name, shape_type)
+            except Exception as e:
+                raise e
             i += 1
-            if not t:
-                return t
+            
             dual_print(f"[{inspect.stack()[0][3]}] - Image {image_name_no_extension} Cropped successfully.")
-        os.remove(full_img_path)
-    return t
+        if(shape_type):
+            os.remove(full_img_path)
 
 
 def crop_files(images_input, folder_name: str, path: str):
@@ -223,10 +228,10 @@ def crop_files(images_input, folder_name: str, path: str):
     """
     for image_name in images_input:
         shape_type = path.split(path_delimiter)[-1]
-        t = process_image(path, folder_name, image_name, shape_type=shape_type)
-        if not t:
-            return
-
+        try:
+            process_image(path, folder_name, image_name, shape_type=shape_type)
+        except Exception as e:
+            raise(e)
 
 def runThreads(input_path: str, folder_name: str, type_):
     """
@@ -281,7 +286,7 @@ def preProcessingMain(input_dir):
             main_folders.append(os.path.join(input_dir, main_folder))
     except FileNotFoundError:
         dual_print(f"[{inspect.stack()[0][3]}] - Input file {INPUT_PATH} not found.", "error")
-        return  # The script can"t run without input
+        raise FileNotFoundError("Input file not found.")
 
     for main_folder in main_folders:  # iterate over the main folders
         for sub_folder in os.listdir(main_folder):
@@ -335,6 +340,3 @@ def main(input_dir, output_dir: str = ""):
     dual_print(
         f"[{inspect.stack()[0][3]}] - {str(total_images_cropped)} Images have been cropped into {str(total_patches_cropped)} Patches.")
     return output_dir
-
-
-main("input")
